1.5.3
-------Cobalt Essentials 1.5.3 Changelog-------
Reworked how groups function in code
Added getGroup method to players
Implemented support for using a playerID in a command by using {<ID>} instead of the username
Overhauled Console output with new `CElog()` function.
Added show debug option to config
Added the ability to color text in the console, assuming the console you are using is compliant with ANSI escape sequences.
Applied the aforementioned colors in a bunch of places.
Fixed bug that caused json files to be corrupted upon shrinking
Added more error-trapping for setgroup and setperm
Implemented security measures to setgroup and setperm to disallow them from granting a permission level higher than the sender's
Groups may now be removed from a player by using setgroup and setting the group to "none"
Moved the default ban reason off default and to guest, where it was supposed to be.